src/requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
backoff==2.2.1
billiard==4.1.0
branca==0.6.0
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.0
Django==4.2.5
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-filter==23.3
django-redis==5.3.0
django-timezone-field==6.0.1
djangorestframework==3.14.0
folium==0.14.0
gunicorn==21.2.0
idna==3.4
Jinja2==3.1.2
kombu==5.3.2
MarkupSafe==2.1.3
numpy==1.25.2
packaging==23.1
prompt-toolkit==3.0.39
psycopg2==2.9.7
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
